Switching numeric answer to int answer. Returning non pointers from database so resolvers are consistent. Adding ability to add meetings and likert answers via the interface.

+++ meeting.go
@@ -4,7 +4,7 @@ import (
 	"time"
 )
 
-const NUMERIC = "numeric"
+const INT = "int"
 
 type Answer struct {
 	QuestionID string      `json:"questionID" bson:"questionID"`
